Moved Dialog to a shared folder and updated imports
